split plan types into modules matching structure in ndc-models
