+++ crates/ndc-query-plan/src/query_plan/mod.rs
@@ -1,0 +1,14 @@
+mod aggregation;
+pub use aggregation::*;
+mod connector_types;
+pub use connector_types::*;
+mod expression;
+pub use expression::*;
+mod fields;
+pub use fields::*;
+mod ordering;
+pub use ordering::*;
+mod requests;
+pub use requests::*;
+mod schema;
+pub use schema::*;
